+++ R/cancensus_cache.R
@@ -1,4 +1,11 @@
-touch_metadata <- function(meta_file,params=NULL){
+metadata_columns <- c("path" , "dataset", "regions", "level", "vectors", "created_at",
+                      "version", "size" ,"last_accessed", "access_count",  "resolution")
+metadata_columns_c <- c("path" , "dataset", "regions", "level", "vectors", "version", "resolution")
+metadata_columns_d <- c("created_at","last_accessed")
+metadata_columns_n <- c("size" , "access_count")
+
+
+generate_metadata <- function(meta_file,params){
   if (file.exists(meta_file)) {
     metadata <- readRDS(meta_file) %>%
       dplyr::mutate(last_accessed=Sys.time())
@@ -32,6 +39,22 @@ touch_metadata <- function(meta_file,params=NULL){
     if (grepl("^CM_geo_",path)) metadata$version="g.1"
     if (grepl("^CM_data_",path)) metadata$version="d.1"
   }
+
+  for (n in setdiff(metadata_columns_c,names(metadata))) {
+    metadata[,n]=NA_character_
+  }
+  # for (n in setdiff(metadata_columns_d,names(metadata))) {
+  #   metadata[,n]=NA
+  # }
+  for (n in setdiff(metadata_columns_n,names(metadata))) {
+    metadata[,n]=NA_real_
+  }
+
+  metadata
+}
+
+touch_metadata <- function(meta_file,params=NULL){
+  metadata <- generate_metadata(meta_file,params)
   saveRDS(metadata,file = meta_file)
 }