+++ R/helpers.R
@@ -84,6 +84,33 @@ as.int = function(x, na.strings = cancensus_na_strings) {
   x
 }
 
+check_recalled_data_and_warn <- function(meta_file,params){
+  cached_data<-generate_metadata(meta_file,params)
+  recalled_data <- list_recalled_cached_data(cached_data,warn_only_once=TRUE)
+  if (!is.null(recalled_data) && nrow(recalled_data)>0) {
+    warning("Currently loaded data has been recalled. Use\nlist_recalled_cached_data()\nto inspect recalled locally cached data and\nremove_recalled_chached_data()\nto remove recalled data.")
+  }
+  d<-NULL
+}
+
+check_for_recalled_data_and_warn <- function(){
+  recalled_data <- list_recalled_cached_data(warn_only_once=TRUE)
+  if (!is.null(recalled_data) && nrow(recalled_data)>0) {
+    warning(paste0("Some locally cached data has been recalled. Use\nlist_recalled_cached_data()\nto inspect recalled locally cached data and\nremove_recalled_chached_data()\nto remove recalled data."))
+  }
+  d<-NULL
+}
+
+first_run_checks <- function(){
+  path=file.path(tempdir(),"cancensus_first_run_checks.info")
+  if (!file.exists(path)) {
+    # first time cancensus is run!
+    check_for_recalled_data_and_warn()
+    readr::write_lines("recall_check\n",path)
+  }
+  d<-NULL
+}
+
 #' A dataset with code table summaries for census data
 #' @name CODES_TABLE
 #' @docType data